Server: reworked file structure, broke createUser into a helper of its own and implemented it in the register function. Made the code much more readable. Created updateCredentials and login. All current endpoints tested and working fully. Now to break login functionality into a helper.
