+++ server/Helpers/authHelpers/register.js
@@ -1,6 +1,9 @@
 const loginModel = require('../../Model/loginModel');
 const userModel = require('../../Model/userModel');
 
+// Secondary Helpers
+const createUser = require('./createUser');
+
 const register = (name,hash,res) => {
     const loginObject = {
         name:name,
@@ -8,38 +11,12 @@ const register = (name,hash,res) => {
     }
     loginModel.findOne({name})
     .then(found => {
-        if (found) {
-            return res.status(400).send('Error');
-        };
-        loginModel.create(loginObject)
-        .then(login => {
-            const userObject = {
-                _id: login._id,
-                data: {
-                    name:name,
-                    teams: {
-                        current: '',
-                        past: []
-                    },
-                    bugs: {
-                        current: '',
-                        past: []
-                    }
-                }
-            }
-            userModel.create(userObject)
-            .then(user => {
-                if (user) {
-                    return res.status(201).send('Success')
-                }
-            })
-            .catch((err) => {
-                return res.status(400).send('Error');
-            })
-        })
-        .catch((err) => {
-            return res.status(400).send('Error');
-        })
+        if (!found) {
+            createUser(loginObject,res)
+        }
+        else {
+            return res.status(400).send('Error')
+        }
     })
     .catch((err) => {
         return res.status(400).send('Error');
